[llms] Keep the 8B decode's embedding table in bf16
The driver promoted embed_tokens to f32, inherited from the 1B/3B-lineage
drivers where the table is small. At 128256x4096 that is a 1 GiB table becoming
2 GiB, on the example that already holds the largest resident BOs. The only use
is a per-token row lookup that is cast straight back to bf16 for the device
input, so the promotion buys nothing -- and the qwen3_8b sibling, the other
K=4096 model, already stores bf16.

Behavior-preserving: bf16 -> f32 -> bf16 and bf16 -> bf16 give the same bits.
Confirmed on NPU2 -- a 64-token greedy generation is byte-identical to the run
before this change, and make verify still passes 2/2.

+++ programming_examples/llms/llama31_8b_q4nx/q4nx_decode_8b.py
@@ -229,7 +229,11 @@ class FusedDecode8B:
         # embed_norm_lmhead()'s tied third return is deliberately not used.
         embed = qm.bf16("model.embed_tokens.weight")
         final_norm = qm.bf16("model.norm.weight")
-        self.embed = np.asarray(embed, np.float32).reshape(self.VOCAB_SIZE, self.K)
+        # Kept in bf16, not promoted to f32 as the 1B/3B-lineage drivers do: the
+        # only use is a per-token row lookup that is cast back to bf16 for the
+        # device input, so f32 would just double a 1 GiB table for nothing. The
+        # qwen3_8b sibling does the same.
+        self.embed = np.asarray(embed, bfloat16).reshape(self.VOCAB_SIZE, self.K)
         final_norm = np.asarray(final_norm, bfloat16)
 
         import decode_dynseq as _dyn
